fixed minimum and maximum to accept doubles instead of uint

+++ JsonSchema.Generation/MaximumAttribute.cs
@@ -13,13 +13,13 @@ namespace Json.Schema.Generation
 		/// <summary>
 		/// The maximum.
 		/// </summary>
-		public uint Value { get; }
+		public decimal Value { get; }
 
 		/// <summary>
 		/// Creates a new <see cref="MaximumAttribute"/> instance.
 		/// </summary>
 		/// <param name="value">The value.</param>
-		public MaximumAttribute(uint value)
+		public MaximumAttribute(decimal value)
 		{
 			Value = value;
 		}
